% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_likelihood_selfimage.R
\name{save_likelihood_selfimage}
\alias{save_likelihood_selfimage}
\title{Save Self-Image Likelihood Rating}
\usage{
save_likelihood_selfimage(selfimage_data, likelihood, file_name)
}
\arguments{
\item{selfimage_data}{A data frame containing self-image data with
columns 'date' and 'likelihood_selfimage'.}

\item{likelihood}{A numeric value between 0 and 100 that indicates how
likely the user thinks their desired positive self-image is to be true.}

\item{file_name}{A character string representing the name of the csv file
where the data should be saved.}
}
\value{
The updated data frame with the new likelihood rating added.
}
\description{
This function saves the daily likelihood rating of the user's
desired positive self-image to a CSV file and updates the data frame.
}
\examples{
# Example usage:
selfimage_data <- data.frame(date = as.Date(character(0)),
likelihood_selfimage = numeric(0))
updated_data <- save_likelihood_selfimage(selfimage_data, 75,
"selfimage_data.csv")
}
