% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_whitebook_entry.R
\name{save_whitebook_entry}
\alias{save_whitebook_entry}
\title{Save Whitebook Entry}
\usage{
save_whitebook_entry(
  whitebook_file,
  event,
  feeling,
  pers_trait,
  wordcloud_file
)
}
\arguments{
\item{whitebook_file}{A character string with the name of the CSV file
of the whitebook data}

\item{event}{A character string with the positive event.}

\item{feeling}{A character string with the feeling associated with the event.}

\item{pers_trait}{A character string with the positive personality trait that
can be associated with the positive event}

\item{wordcloud_file}{A character string with the name of the CSV file
of the wordcloud data}
}
\value{
A list containing the updated whitebook data frame and wordcloud data frame.
}
\description{
This function saves an entry to the whitebook and updates the data frame.
}
\examples{
# Example usage:
updated_data <- save_whitebook_entry("whitebook_data.csv", "Did the dishes", "Proud", "Determined", "wordcloud_data.csv")
print(updated_data)
}
